Enable nvim-lspconfig configuration via lsp.lua
